Big geometry interface rework
BufferType now encapsulates both index and vertex buffers nicely.
Subclasses are now responsible for keeping track of the buffers they need, but there are helper functions now in the form of BufferType constructors.
Removed ALL templating. Can't use it in C++ when source is defined in seperate files to headers. Templating in C++ is nothing but a headache and I wasted the better part of my day discovering this.